fargate/chat_ui/requirements.txt
boto3
langchain==0.0.185
opensearch-py==2.2.0
openapi-schema-pydantic==1.2.4
psutil==5.9.1
pydantic==1.10.7
PyYaml
streamlit==1.22.0
streamlit-ace==0.1.1
streamlit-authenticator==0.2.2
streamlit-chat==0.0.2.2
streamlit-chat-media==0.0.4
transformers==4.29.0
typing-inspect==0.8.0
typing_extensions==4.3.0